requirements.txt
pyaudio==0.2.11
wave==0.0.2
numpy==1.21.3
matplotlib==3.4.3
pyautogui==0.9.53
keyboard==0.13.5
Pillow==8.3.2
requests==2.26.0
opencv-python-headless==4.5.3.56
screeninfo==0.6.7
beautifulsoup4==4.10.0